+++ src/test/integration/uk/gov/hmcts/reform/sandl/snlevents/BaseIntegrationTest.java
@@ -1,10 +1,9 @@
-package integration;
+package uk.gov.hmcts.reform.sandl.snlevents;
 
 import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.hmcts.reform.sandl.snlevents.Application;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = Application.class)
