+++ src/test/integration/uk/gov/hmcts/reform/sandl/snlevents/fakerules/IntegrationTestWithFakeBackend.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.sandl.snlevents.fakerules;
+
+public interface IntegrationTestWithFakeBackend {
+}
